사용하지 않는 기능 코드정리
