--- app/src/main/res/xml/file_paths.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<paths xmlns:android="http://schemas.android.com/apk/res/android">
-    <files-path name="my_images" path="." />
-</paths>
-
-<!--<?xml version="1.0" encoding="utf-8"?>-->
-<!--<paths xmlns:android="http://schemas.android.com/apk/res/android">-->
-<!--    <external-path name="my_images" path="Android/data/com.lee.oneweekonebook/files/Pictures" />-->
-<!--</paths>-->
